refactored functions to get a better overview
