+++ EVM/cli.py
@@ -1,4 +1,3 @@
-#!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 # Volker Göhler 2019,2020
