+++ EVM/filter.py
@@ -1,0 +1,26 @@
+# -*- coding: utf-8 -*-
+# Volker Göhler 2019,2020
+
+import numpy as np
+from scipy import signal as signal, fftpack as fftpack
+
+
+def butter_bandpass_filter(data, lowcut, highcut, fs, order=5):
+    omega = 0.5 * fs
+    low = lowcut / omega
+    high = highcut / omega
+    b, a = signal.butter(order, [low, high], btype='band')
+    y = signal.lfilter(b, a, data, axis=0)
+    return y
+
+
+def temporal_ideal_filter(tensor, low, high, fps, axis=0):
+    fft = fftpack.fft(tensor, axis=axis)
+    frequencies = fftpack.fftfreq(tensor.shape[0], d=1.0 / fps)
+    bound_low = (np.abs(frequencies - low)).argmin()
+    bound_high = (np.abs(frequencies - high)).argmin()
+    fft[:bound_low] = 0
+    fft[bound_high:-bound_high] = 0
+    fft[-bound_low:] = 0
+    iff = fftpack.ifft(fft, axis=axis)
+    return np.abs(iff)
